+++ src/args.cc
@@ -183,7 +183,9 @@ Args::Args(int argc, char **argv)
     } else if (std::strcmp("--align-text-right", argv[0]) == 0) {
       flags.set(15);
     } else if (std::strcmp("--pprompt", argv[0]) == 0) {
-      flags.set(6);
+      PrintHelper::println(stderr,
+                           "WARNING: --pprompt is disabled; its functionality "
+                           "is the default behavior.");
     } else if (std::strncmp("--pfile=", argv[0], 8) == 0) {
       password_file = std::string(argv[0] + 8);
     } else if (std::strcmp("--no-scale-fill", argv[0]) == 0) {
@@ -391,14 +393,6 @@ Args::Args(int argc, char **argv)
     --argc;
     ++argv;
   }
-
-  if (host_ip_addr.empty() && host_unix_socket.empty()) {
-    PrintHelper::println(
-        stderr,
-        "ERROR: --host=<ip_addr> or --host-socket=<path> not specified!");
-    flags.set(0);
-    return;
-  }
 }
 
 void Args::print_usage() {
@@ -419,7 +413,8 @@ void Args::print_usage() {
   PrintHelper::println(
       "  --disable-show-percentage : disable showing song percentage");
   PrintHelper::println("  --align-text-right : Aligns the text to the right");
-  PrintHelper::println("  --pprompt : prompt for password on program start");
+  PrintHelper::println(
+      "  --pprompt : deprecated; it is the default to prompt for password");
   PrintHelper::println(
       "  --pfile=<filename> : get password from specified file");
   PrintHelper::println(
@@ -542,3 +537,12 @@ const std::unique_ptr<Color> &Args::get_text_bg_color() const {
 float Args::get_y_offset() const { return text_y_offset; }
 
 bool Args::is_y_offset_from_top() const { return flags.test(22); }
+
+void Args::set_host_ip_addr(std::string addr) {
+  this->host_ip_addr = std::move(addr);
+}
+
+void Args::set_host_socket(std::string socket) {
+  this->host_unix_socket = std::move(socket);
+  flags.set(23);
+}